Refactor - 변수명 오타 수정, Imorrt 정리
* 변수명 오타 수정
* 사용하지 않는 Import 정리 등

+++ src/main/java/team/snowball/baseball/model/player/PlayerRepository.java
@@ -13,6 +13,8 @@ public interface PlayerRepository {
 
     List<Player> findByTeamId(int id);
 
-    int delete(int id);
+    int delete(Long id);
 
+    //Todo: update 구현
+//    int update(Player player);
 }
